requirements.txt
langchain==0.3.26
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0

langchain-huggingface
langchain-pinecone==0.2.8
langchain-community==0.3.26
langchain-groq==0.3.8

gunicorn==21.2.0
-e .
